Fjalla 1.001 source files
The language support is increased and some technical mistakes have been corrected. for example

• The diacritics are now sourced from 0300 series unicodes with zero width rather than the old default.
• Legacy diacritics remain with some spacing in case they need to be referenced in text.
• Vertical metrics have ben made consistent with current Google practice to avoid clipping.
• OpenType naming tables are added and have NID 16 & 17 entries so that Font Menus in Adobe apps say Fjalla with a style of "One"

If this is mastered into TTFs etc & TTFA autohinted witha contempoary version it will be bit of an improvement.

It still needs some symbols to be added ( partial diff etc ) to be 100% of mac roman etc.